Merged/moved my RecyleBin module into Defte's existing file. I removed the registry key logic to fetch the username since it's already given in the metadata file ($I).
